learnwt/requirements.txt
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
black==19.3b0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
decorator==4.4.0
Django==2.2.3
django-debug-toolbar==2.0
django-extensions==2.1.9
django-materialize-form==0.1.1
django-materializecss-form==1.1.14
django-modelcluster==4.4
django-recaptcha==2.0.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.6
html5lib==1.0.1
idna==2.8
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
Unidecode==1.1.1
urllib3==1.25.3
wagtail==2.6
wagtail-django-recaptcha==1.0
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
